Add conditional import for OTPForm
On projectis that dont use OTP, we dont import the OTPForm because it
brings a dependency to django-otp.
